+++ cached_network_image_platform_interface/lib/cached_network_image_platform_interface.dart
@@ -23,25 +23,6 @@ class ImageLoader {
   /// loads the images async and gives the resulted codecs on a Stream. The
   /// Stream gives the option to show multiple images after each other.
   @Deprecated('Use loadImageAsync instead')
-  Stream<ui.Codec> loadAsync(
-    String url,
-    String? cacheKey,
-    StreamController<ImageChunkEvent> chunkEvents,
-    DecoderCallback decode,
-    BaseCacheManager cacheManager,
-    int? maxHeight,
-    int? maxWidth,
-    Map<String, String>? headers,
-    ValueChanged<Object>? errorListener,
-    ImageRenderMethodForWeb imageRenderMethodForWeb,
-    VoidCallback evictImage,
-  ) {
-    throw UnimplementedError();
-  }
-
-  /// loads the images async and gives the resulted codecs on a Stream. The
-  /// Stream gives the option to show multiple images after each other.
-  @Deprecated('Use loadImageAsync instead')
   Stream<ui.Codec> loadBufferAsync(
     String url,
     String? cacheKey,