+++ cached_network_image/CHANGELOG.md
@@ -4,7 +4,8 @@
 * Specify types
 * Add error to ErrorListener
 * Update example
-* Use `loadImage` instead `loadBuffer` and `load`
+* Use `loadImage` instead `loadBuffer`
+* Remove [`load`](https://github.com/flutter/flutter/pull/132679)
 
 ## [3.2.3] - 2022-11-25
 * Correctly dispose image stream handler